requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
gspread>=6.1.2
google-auth>=2.27.0
groq>=0.4.2
httpx>=0.26.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
